@@ -3,38 +3,47 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev",
+    "dev": "next dev --turbopack",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "eslint ."
   },
   "dependencies": {
-    "@cloudinary/react": "^1.11.2",
-    "@cloudinary/url-gen": "^1.11.1",
-    "@vercel/analytics": "^1.0.2",
-    "axios": "^1.6.8",
-    "bootstrap": "^5.2.0",
-    "chart.js": "^4.3.0",
-    "cloudinary": "^1.40.0",
-    "dayjs": "^1.11.1",
-    "mongodb": "^4.17.2",
-    "next": "^14.2.3",
-    "next-auth": "^4.23.0",
-    "react": "18.0.0",
-    "react-bootstrap": "^2.5.0",
-    "react-chartjs-2": "^5.2.0",
-    "react-datepicker": "^4.7.0",
-    "react-dom": "18.0.0",
-    "react-table": "^7.7.0",
-    "react-toastify": "^9.0.8"
+    "@cloudinary/react": "^1.14.3",
+    "@cloudinary/url-gen": "^1.22.0",
+    "@vercel/analytics": "^1.5.0",
+    "bootstrap": "^5.3.8",
+    "chart.js": "^4.5.1",
+    "cloudinary": "^2.8.0",
+    "dayjs": "^1.11.18",
+    "lodash": "^4.17.21",
+    "mongodb": "^6.20.0",
+    "next": "15.5.7",
+    "next-auth": "^5.0.0-beta.29",
+    "react": "19.2.0",
+    "react-bootstrap": "^2.10.10",
+    "react-chartjs-2": "^5.3.0",
+    "react-datepicker": "^8.8.0",
+    "react-dom": "19.2.0",
+    "react-select": "^5.10.2",
+    "react-table": "^7.8.0",
+    "react-toastify": "^11.0.5"
   },
   "devDependencies": {
-    "@types/node": "18.17.1",
-    "@types/react": "18.0.0",
-    "@types/react-dom": "18.0.0",
-    "@types/react-table": "^7.7.11",
-    "eslint": "8.13.0",
-    "eslint-config-next": "12.1.4",
-    "typescript": "4.6.3"
+    "@eslint/js": "^9.38.0",
+    "@playwright/test": "^1.56.1",
+    "@types/node": "24.9.1",
+    "@types/react": "19.2.2",
+    "@types/react-dom": "19.2.2",
+    "@types/react-table": "^7.7.20",
+    "eslint": "^9.38.0",
+    "eslint-config-next": "^16.0.0",
+    "globals": "^16.4.0",
+    "typescript": "5.9.3"
+  },
+  "overrides": {
+    "@types/react": "19.2.2",
+    "@types/react-dom": "19.2.2",
+    "react": "19.2.0"
   }
 }
