@@ -1,5 +1,5 @@
 type Props = {
-  type: 'youtube' | 'twitch' | 'twitter' | 'discord' | 'bsky'
+  type: 'youtube' | 'twitch' | 'bluesky' | 'discord'
   size?: number
   className?: string
 }
@@ -29,7 +29,7 @@ export const Icon = (props: Props) => {
         width={size}
         height={size}
         fill='currentColor'
-        className={`bi bi-twitch ${className}`}
+        className={`bi bi-youtube ${className}`}
         viewBox='0 0 16 16'
       >
         <path d='M3.857 0 1 2.857v10.286h3.429V16l2.857-2.857H9.57L14.714 8V0H3.857zm9.714 7.429-2.285 2.285H9l-2 2v-2H4.429V1.143h9.142v6.286z' />
@@ -38,17 +38,17 @@ export const Icon = (props: Props) => {
     )
   }
 
-  if (type === 'twitter' || type === 'bsky') {
+  if (type === 'bluesky') {
     return (
       <svg
         xmlns='http://www.w3.org/2000/svg'
         width={size}
         height={size}
         fill='currentColor'
-        className={`bi bi-twitter ${className}`}
-        viewBox='0 0 16 16'
+        className={`bi bi-bluesky ${className}`}
+        viewBox='0 0 24 24'
       >
-        <path d='M5.026 15c6.038 0 9.341-5.003 9.341-9.334 0-.14 0-.282-.006-.422A6.685 6.685 0 0 0 16 3.542a6.658 6.658 0 0 1-1.889.518 3.301 3.301 0 0 0 1.447-1.817 6.533 6.533 0 0 1-2.087.793A3.286 3.286 0 0 0 7.875 6.03a9.325 9.325 0 0 1-6.767-3.429 3.289 3.289 0 0 0 1.018 4.382A3.323 3.323 0 0 1 .64 6.575v.045a3.288 3.288 0 0 0 2.632 3.218 3.203 3.203 0 0 1-.865.115 3.23 3.23 0 0 1-.614-.057 3.283 3.283 0 0 0 3.067 2.277A6.588 6.588 0 0 1 .78 13.58a6.32 6.32 0 0 1-.78-.045A9.344 9.344 0 0 0 5.026 15z' />
+        <path d="M12 10.8c-1.087-2.114-4.046-6.053-6.798-7.995C2.566.944 1.561 1.266.902 1.565C.139 1.908 0 3.08 0 3.768c0 .69.378 5.65.624 6.479c.815 2.736 3.713 3.66 6.383 3.364c.136-.02.275-.039.415-.056c-.138.022-.276.04-.415.056c-3.912.58-7.387 2.005-2.83 7.078c5.013 5.19 6.87-1.113 7.823-4.308c.953 3.195 2.05 9.271 7.733 4.308c4.267-4.308 1.172-6.498-2.74-7.078a8.741 8.741 0 0 1-.415-.056c.14.017.279.036.415.056c2.67.297 5.568-.628 6.383-3.364c.246-.828.624-5.79.624-6.478c0-.69-.139-1.861-.902-2.206c-.659-.298-1.664-.62-4.3 1.24C16.046 4.748 13.087 8.687 12 10.8"/>
       </svg>
     )
   }
