@@ -1,7 +1,9 @@
-import axios from 'axios'
+'use client'
+
 import React, { useState } from 'react'
 import { Modal } from 'react-bootstrap'
 import { toast } from 'react-toastify'
+import { updateVodsAction } from 'app/actions'
 
 type Props = {
   game: Game
@@ -13,17 +15,16 @@ export const VodDialog = (props: Props) => {
   const [show, setShow] = useState(false)
   const [textValue, setTextValue] = useState(game.vods?.join('\n') || '')
 
-  const updateVods = () => {
-    const vods = textValue ? textValue.split('\n') : null
-    axios
-      .put('api/vods', { id: game._id, vods })
-      .then((_res) => {
-        toast.success('VODS saved 👌')
-      })
-      .catch((err) => {
-        toast.error('Save failed')
-        console.log('ERROR: ', err)
-      })
+  const updateVods = async () => {
+    const formData = new FormData()
+    formData.append('id', game._id!)
+    formData.append('vods', textValue)
+    var res = await updateVodsAction(formData)
+    if (res) {
+      toast.success('Vods updated')
+    } else {
+      toast.error('Failed to update vods')
+    }
   }
 
   return (
@@ -31,12 +32,12 @@ export const VodDialog = (props: Props) => {
       <a href='#' onClick={() => setShow(true)}>
         Add vods
       </a>
-      <Modal show={show} onHide={() => setShow(false)} centered>
+      <Modal show={show} onHide={() => setShow(false)} centered id="vodDialog">
         <Modal.Header closeButton>
           <Modal.Title>Add vods</Modal.Title>
         </Modal.Header>
         <Modal.Body>
-          <textarea cols={45} rows={5} value={textValue} onChange={(e) => setTextValue(e.target.value)} />
+          <textarea cols={45} rows={5} id="vodsArea" value={textValue} onChange={(e) => setTextValue(e.target.value)} />
         </Modal.Body>
         <Modal.Footer>
           <button className='btn btn-light' onClick={() => updateVods()}>
