@@ -10,23 +10,26 @@ type Game = {
   igdbId: number
   igdbUrl: string
   keywords: Array<string>
-  notPollable: string | null
   platform: string | null
   rating: number | null
   releaseYear: number | null
   timeSpent: number | null
-  tss: boolean | null
   streamed: boolean | null
   vods: Array<string> | null
-}
-
-type RandomGame = Game & {
-  selected?: number
+  hltbMain: number | null
+  hltbExtra: number | null
+  hltbCompletionist: number | null
+  tags: Array<string> | null
+  additionalTimeSpent: number | null
+  addedDate: string | null // not nullable after db update
 }
 
 type GameOptions = {
   id: number
-  title: string
-  year: number
+  name: string
+  release_dates: Array<{
+    y: number
+    id: number
+  }>
   url: string
 }
