@@ -1,8 +1,9 @@
-import { AdminCell, CheckmarkCell, CommentCell, DateCell, FinishedCell, TitleCell, VodCell } from '../components/Cells'
+import { Column } from 'react-table'
+import { AdminCell, CommentCell, DateCell, FinishedCell, TagCell, TitleCell, VodCell } from '../components/Cells'
 import { ScoreIndicator } from '../components/ScoreIndicator'
-import { dateSort, scoreSort } from './utils'
+import { dateSort, scoreSort, titleSort } from './utils'
 
-export const gameTableColumns = [
+export const gameTableColumns: Column<Partial<Game>>[] = [
   {
     Header: 'Date',
     accessor: 'finishedDate',
@@ -14,20 +15,22 @@ export const gameTableColumns = [
     Header: 'Game',
     accessor: 'title',
     Cell: ({ value, row, showCovers }) => TitleCell({ value, row, showCovers }),
+    sortType: titleSort,
+    sortDescFirst: true
   },
   {
     Header: 'Rating',
     accessor: 'rating',
     sortDescFirst: true,
     Cell: ({ value }) => {
-      return <ScoreIndicator rating={value} />
+      return <ScoreIndicator rating={(value as number | null)} />
     },
     sortType: scoreSort,
   },
   {
     Header: 'Comments',
     accessor: 'comment',
-    Cell: ({ value, row }) => CommentCell({ value, row }),
+    Cell: ({ value, row, handleTagFilterChange }) => CommentCell({ value, row, handleTagFilterChange }),
     disableSortBy: true,
   },
   {
@@ -50,21 +53,28 @@ export const gameTableColumns = [
   },
 ]
 
-export const backlogTableColumns = [
+export const backlogTableColumns: Column<Partial<Game & { hltbString: string }>>[] = [
   {
     Header: 'Game',
     accessor: 'title',
-    Cell: TitleCell,
+    Cell: ({ value, row }) => TitleCell({ value, row, showCovers: false }),
+    sortType: titleSort,
+    sortDescFirst: true
   },
   {
-    Header: 'Blocked by',
-    accessor: 'notPollable',
+    Header: 'Tags',
+    accessor: 'tags',
+    Cell: ({ value, row, handleTagFilterChange }) => TagCell({ value, row, handleTagFilterChange }),
+  },
+  {
+    Header: 'Howlongtobeat',
+    accessor: 'hltbString',
   },
   {
     Header: 'Admin',
     accessor: '_id',
     disableGlobalFilter: true,
     disableSortBy: true,
-    Cell: ({ value, row }) => AdminCell({ value, row, showNextButton: true }),
+    Cell: ({ value, row }) => AdminCell({ value, row }),
   },
 ]
