@@ -1,5 +1,28 @@
 import { Cell, ColumnInstance, Row } from 'react-table'
 
+export const getHltbString = (game: Game) => {
+  const main = game.hltbMain || 0
+  const extra = game.hltbExtra || 0
+  const completionist = game.hltbCompletionist || 0
+
+  if (!main) {
+    return '-'
+  }
+  if (main === extra || main === completionist) {
+    return `${main} h`
+  }
+  if (extra > 0) {
+    return `${main}-${extra} h`
+  }
+  if (completionist > 0) {
+    return `${main}-${completionist} h`
+  }
+  if (extra === 0 && completionist === 0) {
+    return `${main} h`
+  }
+  return ''
+}
+
 /*********************************************************************
  * SORTING FUNCTIONS
  *********************************************************************/
@@ -22,6 +45,23 @@ export const scoreSort = (rowA, rowB, id) => {
   return rowA.values[id] - rowB.values[id]
 }
 
+export const titleSort = (rowA, rowB, id) => {
+  const rowAValue: string = rowA.values[id]
+  const rowBValue: string = rowB.values[id]
+
+  const rowAFormatted = rowAValue.startsWith("The ") ? rowAValue.substring(4) : rowAValue
+  const rowBFormatted = rowBValue.startsWith("The ") ? rowBValue.substring(4) : rowBValue
+
+  return rowBFormatted.localeCompare(rowAFormatted)
+}
+
+export const titleSortSimple = (valueA: string, valueB: string) => {
+  const valueAFormatted = valueA.startsWith("The ") ? valueA.substring(4) : valueA
+  const valueBFormatted = valueB.startsWith("The ") ? valueB.substring(4) : valueB
+
+  return valueAFormatted.localeCompare(valueBFormatted)
+}
+
 /*********************************************************************
  * COLUMN FORMATTING
  *********************************************************************/
@@ -34,25 +74,26 @@ export const formatCell = (cell: Cell<object, any>, row: Row<object>, props?: an
   const { column } = cell
   const columnId = column.id
   const showCovers = props?.showCovers
+  const handleTagFilterChange = props?.handleTagFilterChange
 
   switch (true) {
     case COLUMN_IDS.TITLE.includes(columnId):
-      return renderCellWithProps(cell, { key: columnId + row.id }, { showCovers })
+      return renderCellWithProps(cell, { key: columnId + row.id }, { showCovers, handleTagFilterChange })
 
     case COLUMN_IDS.CENTERED.includes(columnId):
       return renderCellWithProps(cell, {
         key: columnId + row.id,
         style: { textAlign: 'center' },
-      })
+      }, { handleTagFilterChange })
 
     default:
-      return renderCellWithProps(cell, { key: columnId + row.id })
+      return renderCellWithProps(cell, { key: columnId + row.id }, { handleTagFilterChange })
   }
 }
 
 const renderCellWithProps = (cell: Cell<object, any>, attrs: object, props?: object) => {
   return (
-    <td {...cell.getCellProps()} {...attrs}>
+    <td  {...cell.getCellProps()} {...attrs} key={attrs['key']}>
       {cell.render('Cell', props)}
     </td>
   )
